% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kform_verification.R
\name{myKFunction}
\alias{myKFunction}
\title{k-Form function}
\usage{
myKFunction(input1, input2, input3 = NULL, input4 = NULL, input5)
}
\arguments{
\item{input1}{Mean value}

\item{input2}{Standard deviation}

\item{input3}{LSL lower specification limit}

\item{input4}{USL upper specification limit}

\item{input5}{kFac,  tabled acceptance criterion}
}
\value{
Result String
}
\description{
k-Form function
}
\examples{

# Example ISO 3951-2 Chapter 16.2
 myKFunction(6.551,0.3251,4.0,NA,2.580)

# Example ISO 3951-2 Chapter 16.2
 myKFunction(54.615385,3.3301267,NA,60,1.426)



}
