% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pform_verification.R
\name{myPFunction}
\alias{myPFunction}
\title{p-Form function}
\usage{
myPFunction(input1, input2, input3, input4, input5, input6, input7)
}
\arguments{
\item{input1}{Mean value}

\item{input2}{Standard deviation}

\item{input3}{LSL lower specification limit}

\item{input4}{USL upper specification limit}

\item{input5}{p*,  tabled acceptance criterion}

\item{input6}{fs, tabled factor to calculate S_max}

\item{input7}{n, sample size}
}
\value{
Result String
}
\description{
p-Form function
}
\examples{

# Example ISO 3951-2 Chapter 16.3.2.4
 myPFunction(64.223,2.7899,60,70,0.051590,0.274,13)

# Example ISO 3951-2 Chapter 16.3.2.5
 myPFunction(64.223,2.7899,60,70,0.051590,0.285,13)


}
